refactor: 요금, 경로 패키지 분리
