+++ src/main/java/subway/path/domain/fare/DistanceFare.java
@@ -1,4 +1,4 @@
-package subway.path.domain;
+package subway.path.domain.fare;
 
 public class DistanceFare {
