+++ src/main/java/subway/path/domain/fare/LoginMemberAgeFare.java
@@ -1,4 +1,4 @@
-package subway.path.domain;
+package subway.path.domain.fare;
 
 import subway.member.domain.LoginMember;
