+++ src/main/java/subway/path/domain/path/DistanceLineEdge.java
@@ -1,4 +1,4 @@
-package subway.path.domain;
+package subway.path.domain.path;
 
 import org.jgrapht.graph.DefaultWeightedEdge;
 import subway.line.domain.Line;
