+++ src/main/java/subway/path/domain/path/PathValue.java
@@ -1,4 +1,4 @@
-package subway.path.domain;
+package subway.path.domain.path;
 
 import subway.station.domain.Station;
