+++ src/test/java/subway/path/domain/LineFareTest.java
@@ -4,6 +4,7 @@ import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
 import subway.line.domain.Line;
+import subway.path.domain.fare.LineFare;
 
 import java.util.Arrays;
 import java.util.List;
